Extract Handlers module for handler factory methods
